requirements.txt
anthropic==0.72.0
docx==0.2.4
langchain_chroma==1.0.0
langchain_community==0.4.1
langchain_core==1.0.3
langchain_google_genai==3.0.1
langchain_openai==1.0.2
langchain_text_splitters==1.0.0
openai==2.7.1
openpyxl==3.1.5
pandas==2.3.3
protobuf==6.33.0
py4j==0.10.9.9
pyelftools==0.32
python-dotenv==1.2.1
tqdm==4.67.1
jq==1.10.0
dtw-python==1.7.2
scipy==1.16.3
matplotlib==3.10.8